# PERF: use GEOSCoordSeq_copyToBuffer for get_coordinates (GEOS >= 3.10)