Similar as https://github.com/pygeos/pygeos/pull/436, but now for getting coordinates.

Small example showing the speed-up (10,000 polygons of 100 vertices):

```python
geoms = shapely.polygons(np.random.randn(10000, 100, 2))

In [10]: %timeit shapely.get_coordinates(l)
13.6 ms ± 314 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # <-- main
6.46 ms ± 322 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # <-- PR
```